Django-FastApi-Integration/requirements.txt
# python == 3.14.2
Django == 5.2
fastapi == 0.115.0
uvicorn == 0.30.0
pydantic == 2.12.5
httpx == 0.28.1
python-multipart == 0.0.21
sqlalchemy == 2.0.45
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0